# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminLoginInput {
  email: string;
  password: string;
}

export interface AdminSession {
  email: string;
}

export interface ChangePasswordInput {
  currentPassword: string;
  /** @minLength 8 */
  newPassword: string;
}

export interface AdminProduct {
  id: number;
  subcategoryDbId: number;
  name: string;
  shortName: string;
  description: string;
  priceLbp: number;
  priceUsd: string;
  /** @nullable */
  imageUrl?: string | null;
  recipe: string;
  flavors: string[];
  sortOrder: number;
  hidden: boolean;
  soldOut: boolean;
  deleted: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface AdminSubcategory {
  id: number;
  sectionSlug: string;
  subcategoryId: string;
  name: string;
  description: string;
  themeColor: string;
  accentColor: string;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder: number;
  hidden: boolean;
  deleted: boolean;
  products: AdminProduct[];
  createdAt: string;
  updatedAt: string;
}

export type AdminSectionTheme = { [key: string]: unknown };

export interface AdminSection {
  id: number;
  slug: string;
  name: string;
  subtitle: string;
  sortOrder: number;
  hidden: boolean;
  theme: AdminSectionTheme;
  subcategories: AdminSubcategory[];
}

export interface AdminRelease {
  id: number;
  version: number;
  label: string;
  publishedBy: string;
  isCurrent: boolean;
  publishedAt: string;
}

export interface MediaAsset {
  id: number;
  url: string;
  altText: string;
  filename: string;
  createdAt: string;
}

export interface ExchangeRate {
  /** LBP per 1 USD (e.g. 89500) */
  ratePerUsd: number;
  /** Round up to nearest N LBP (e.g. 50000) */
  roundingTo: number;
}

export type ContentSnapshotSettings = {[key: string]: string};

export interface ContentSnapshot {
  sections: AdminSection[];
  settings: ContentSnapshotSettings;
  publishedAt: string;
}

export type SectionUpdateTheme = { [key: string]: unknown };

export interface SectionUpdate {
  name?: string;
  subtitle?: string;
  hidden?: boolean;
  theme?: SectionUpdateTheme;
}

export interface SubcategoryInput {
  subcategoryId?: string;
  name: string;
  description?: string;
  themeColor?: string;
  accentColor?: string;
  imageUrl?: string;
}

export interface SubcategoryUpdate {
  name?: string;
  description?: string;
  themeColor?: string;
  accentColor?: string;
  /** @nullable */
  imageUrl?: string | null;
  hidden?: boolean;
  sortOrder?: number;
}

export interface ProductInput {
  name: string;
  shortName?: string;
  description?: string;
  priceLbp: number;
  imageUrl?: string;
  recipe?: string;
  flavors?: string[];
}

export interface ProductUpdate {
  name?: string;
  shortName?: string;
  description?: string;
  priceLbp?: number;
  /** @nullable */
  imageUrl?: string | null;
  recipe?: string;
  flavors?: string[];
  hidden?: boolean;
  soldOut?: boolean;
  sortOrder?: number;
}

export interface ReorderInput {
  ids: number[];
}

export interface PublishInput {
  label?: string;
}

export interface SettingsUpdate {[key: string]: string}

export interface ExchangeRateUpdate {
  ratePerUsd: number;
  roundingTo?: number;
}

export type GetAdminSettings200 = {[key: string]: string};

export type UpdateAdminSettings200 = {[key: string]: string};