# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public/content:
    get:
      operationId: getPublishedContent
      tags: [public-content]
      summary: Get currently published content
      responses:
        "200":
          description: Published content snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentSnapshot"
        "404":
          description: No published content yet

  # ─── Auth ─────────────────────────────────────────────────────────────────────
  /admin/auth/login:
    post:
      operationId: adminLogin
      tags: [admin-auth]
      summary: Log in as admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
...
  /admin/auth/session:
    get:
      operationId: getAdminSession
      tags: [admin-auth]
      summary: Get current session
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated
  /admin/auth/change-password:
    post:
      operationId: adminChangePassword
      tags: [admin-auth]
      summary: Change own password (must be authenticated)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "204":
          description: Password changed
        "400":
          description: Invalid input
        "401":
          description: Not authenticated or wrong current password

  # ─── CMS sections ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AdminLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AdminSession:
      type: object
      required: [email]
      properties:
        email:
          type: string
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8

    # ─── Content shapes ──────────────────────────────────────────────────────────
    AdminProduct:
      type: object
      required: [id, subcategoryDbId, name, shortName, description, priceLbp, priceUsd, recipe, flavors, sortOrder, hidden, soldOut, deleted, createdAt, updatedAt]
      properties:
        id:
          type: integer
        subcategoryDbId:
          type: integer
        name:
          type: string
        shortName:
          type: string
        description:
          type: string
        priceLbp:
          type: integer
        priceUsd:
          type: string
        imageUrl:
          type: ["string", "null"]
        recipe:
          type: string
        flavors:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
        hidden:
          type: boolean
        soldOut:
          type: boolean
        deleted:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    AdminSubcategory:
      type: object
      required: [id, sectionSlug, subcategoryId, name, description, themeColor, accentColor, sortOrder, hidden, deleted, products, createdAt, updatedAt]
      properties:
        id:
          type: integer
        sectionSlug:
          type: string
        subcategoryId:
          type: string
        name:
          type: string
        description:
          type: string
        themeColor:
          type: string
        accentColor:
          type: string
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        hidden:
          type: boolean
        deleted:
          type: boolean
        products:
          type: array
          items:
            $ref: "#/components/schemas/AdminProduct"
        createdAt:
          type: string
        updatedAt:
          type: string
    AdminSection:
      type: object
      required: [id, slug, name, subtitle, sortOrder, hidden, theme, subcategories]
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        subtitle:
          type: string
        sortOrder:
          type: integer
        hidden:
          type: boolean
        theme:
          type: object
          additionalProperties: true
        subcategories:
          type: array
          items:
            $ref: "#/components/schemas/AdminSubcategory"
...
    ContentSnapshot:
      type: object
      required: [sections, settings, publishedAt]
      properties:
        sections:
          type: array
          items:
            $ref: "#/components/schemas/AdminSection"
        settings:
          type: object
          additionalProperties:
            type: string
        publishedAt:
          type: string

    # ─── Input/Update schemas ─────────────────────────────────────────────────────
...