# lib/api-zod/src/generated/types/adminSubcategory.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminProduct } from './adminProduct';

export interface AdminSubcategory {
  id: number;
  sectionSlug: string;
  subcategoryId: string;
  name: string;
  description: string;
  themeColor: string;
  accentColor: string;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder: number;
  hidden: boolean;
  deleted: boolean;
  products: AdminProduct[];
  createdAt: string;
  updatedAt: string;
}